Move unicode test to unit tests.
