--- test/UnicodeTest/vc2012/UnicodeTest.sln
@@ -1,20 +1,0 @@
-
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Express 2012 for Windows Desktop
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "UnicodeTest", "UnicodeTest.vcxproj", "{0D045A9C-B68D-45AE-84E5-3097B5212BAD}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Win32 = Debug|Win32
-		Release|Win32 = Release|Win32
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{0D045A9C-B68D-45AE-84E5-3097B5212BAD}.Debug|Win32.ActiveCfg = Debug|Win32
-		{0D045A9C-B68D-45AE-84E5-3097B5212BAD}.Debug|Win32.Build.0 = Debug|Win32
-		{0D045A9C-B68D-45AE-84E5-3097B5212BAD}.Release|Win32.ActiveCfg = Release|Win32
-		{0D045A9C-B68D-45AE-84E5-3097B5212BAD}.Release|Win32.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-EndGlobal
