--- test/UnicodeTest/xcode/UnicodeTest_Prefix.pch
@@ -1,13 +1,0 @@
-
-#if defined( __cplusplus )
-	#include "cinder/Cinder.h"
-	
-	#include "cinder/app/App.h"
-	
-	#include "cinder/gl/gl.h"
-	
-	#include "cinder/CinderMath.h"
-	#include "cinder/Matrix.h"
-	#include "cinder/Vector.h"
-	#include "cinder/Quaternion.h"
-#endif
